chore: rename internal/relay
