+++ src/core/Relay.ts
@@ -7,8 +7,8 @@ import * as Signature from 'ox/Signature'
 import type { Client } from 'viem'
 import * as Account from './Account.js'
 import type { Chain } from './Chains.js'
-import type * as Capabilities from './internal/relay/typebox/capabilities.js'
-import type * as Quote from './internal/relay/typebox/quote.js'
+import type * as Capabilities from './internal/rpcServer/typebox/capabilities.js'
+import type * as Quote from './internal/rpcServer/typebox/quote.js'
 import type { MaybePromise, OneOf, RequiredBy } from './internal/types.js'
 import * as Actions from './internal/viem/actions.js'
 import * as Key from './Key.js'
