+++ src/core/internal/porto.ts
@@ -17,7 +17,7 @@ import type * as Mode from '../Mode.js'
 import type { Config, Store } from '../Porto.js'
 import type * as RpcSchema from '../RpcSchema.js'
 import type * as Provider from './provider.js'
-import type * as RpcSchema_relay from './relay/rpcSchema.js'
+import type * as RpcSchema_relay from './rpcServer/rpcSchema.js'
 
 export type Client<chain extends Chains.Chain = Chains.Chain> = viem_Client<
   Transport,
