+++ src/core/internal/rpcServer/rpcSchema.ts
@@ -1,5 +1,5 @@
 /**
- * JSON-RPC Schema for Relay.
+ * JSON-RPC Schema.
  *
  * @see https://github.com/ithacaxyz/relay/blob/77d1e54e3c7b7268d4e9e9bd89a42637125d9b89/src/rpc.rs#L59-L142
  */
