+++ src/core/internal/viem/actions.test.ts
@@ -8,7 +8,7 @@ import { exp1Abi, exp1Address } from '../../../../test/src/porto.js'
 import { getPorto } from '../../../../test/src/porto.js'
 import * as Key from '../../Key.js'
 import { sendCalls } from '../../Relay.js'
-import type * as Capabilities from '../relay/typebox/capabilities.js'
+import type * as Capabilities from '../rpcServer/typebox/capabilities.js'
 import {
   createAccount,
   getAccounts,
