+++ src/core/internal/viem/actions.ts
@@ -28,8 +28,8 @@ import {
   getExecuteError,
 } from 'viem/experimental/erc7821'
 import type { sendCalls } from '../../Relay.js'
-import type * as RpcSchema from '../relay/rpcSchema.js'
-import * as Rpc from '../relay/typebox/rpc.js'
+import type * as RpcSchema from '../rpcServer/rpcSchema.js'
+import * as Rpc from '../rpcServer/typebox/rpc.js'
 import * as Schema from '../typebox/schema.js'
 import { Value } from '../typebox/schema.js'
